Moving 'rinha' for it's own directory and changing some names
